Replaced float64 with float

+++ oemer/bbox.py
@@ -6,7 +6,6 @@ from numpy import ndarray
 from typing import List
 from typing import Tuple
 from typing import Union, Any
-from numpy import float64
 
 BBox = Tuple[int, int, int, int]
 
@@ -35,7 +34,7 @@ def get_edge(data):
     return data
 
 
-def merge_nearby_bbox(bboxes: List[BBox], distance: float64, x_factor: int = 1, y_factor: int = 1) -> List[BBox]:
+def merge_nearby_bbox(bboxes: List[BBox], distance: float, x_factor: int = 1, y_factor: int = 1) -> List[BBox]:
     model = AgglomerativeClustering(n_clusters=None, distance_threshold=distance, compute_full_tree=True)
     centers = np.array([(bb[0]+bb[2], bb[1]+bb[3]) for bb in bboxes]) / 2
     centers[:, 0] *= x_factor  # Increase/decrease the x distance
